# Aplicação/backend/static_docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PM AI MVP API",
    "description": "\n    ## PM AI MVP - Sistema de Gestão de Projetos com IA\n    \n    API completa para gestão de projetos com funcionalidades avançadas de IA.\n    \n    ### Funcionalidades Principais:\n    \n    * **Autenticação**: Sistema de autenticação OAuth com Google\n    * **Projetos**: Gestão completa de projetos e portfólios\n    * **Equipes**: Gestão de membros da equipe e funções\n    * **Clientes**: Gestão de clientes e níveis de comunicação\n    * **Riscos**: Análise e gestão de riscos com IA\n    * **Checklists**: Sistema de checklists para projetos\n    * **Action Items**: Gestão de itens de ação e tarefas\n    \n    ### Arquitetura:\n    \n    * **Services Layer**: Lógica de negócio separada dos controllers\n    * **Repository Pattern**: Camada de abstração para acesso a dados\n    * **Cache Redis**: Sistema de cache para performance\n    * **Processamento Assíncrono**: Tarefas em background com Celery\n    * **IA Integration**: Integração com Gemini API para análise de riscos\n    \n    ### Tecnologias:\n    \n    * **FastAPI**: Framework web moderno e rápido\n    * **SQLAlchemy**: ORM robusto com suporte assíncrono\n    * **PostgreSQL**: Banco de dados relacional\n    * **Redis**: Sistema de cache\n    * **Celery**: Processamento assíncrono\n    * **Pydantic**: Validação de dados\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/team-members/": {
      "post": {
        "tags": [
          "team-members"
        ],
        "summary": "Create Team Member",
        "description": "Create a new team member.",
        "operationId": "create_team_member_api_v1_team_members__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "team-members"
        ],
        "summary": "Get Team Members",
        "description": "Get all team members for a project.",
        "operationId": "get_team_members_api_v1_team_members__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Project ID",
              "title": "Project Id"
            },
            "description": "Project ID"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "role_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by role",
              "title": "Role Filter"
            },
            "description": "Filter by role"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberResponse"
                  },
                  "title": "Response Get Team Members Api V1 Team Members  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team-members/{team_member_id}": {
      "get": {
        "tags": [
          "team-members"
        ],
        "summary": "Get Team Member",
        "description": "Get a specific team member by ID.",
        "operationId": "get_team_member_api_v1_team_members__team_member_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "team-members"
        ],
        "summary": "Update Team Member",
        "description": "Update a team member.",
        "operationId": "update_team_member_api_v1_team_members__team_member_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "team-members"
        ],
        "summary": "Delete Team Member",
        "description": "Delete a team member.",
        "operationId": "delete_team_member_api_v1_team_members__team_member_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Member Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team-members/project/{project_id}/statistics": {
      "get": {
        "tags": [
          "team-members"
        ],
        "summary": "Get Team Statistics",
        "description": "Get team statistics for a project.",
        "operationId": "get_team_statistics_api_v1_team_members_project__project_id__statistics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team-members/project/{project_id}/by-role": {
      "get": {
        "tags": [
          "team-members"
        ],
        "summary": "Get Team Members By Role",
        "description": "Get team members grouped by role.",
        "operationId": "get_team_members_by_role_api_v1_team_members_project__project_id__by_role_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team-members/{team_member_id}/activate": {
      "post": {
        "tags": [
          "team-members"
        ],
        "summary": "Activate Team Member",
        "description": "Activate a team member.",
        "operationId": "activate_team_member_api_v1_team_members__team_member_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team-members/{team_member_id}/deactivate": {
      "post": {
        "tags": [
          "team-members"
        ],
        "summary": "Deactivate Team Member",
        "description": "Deactivate a team member.",
        "operationId": "deactivate_team_member_api_v1_team_members__team_member_id__deactivate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TeamMemberCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "project_id",
          "role"
        ],
        "title": "TeamMemberCreate"
      },
      "TeamMemberResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "project_id",
          "role",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "TeamMemberResponse"
      },
      "TeamMemberUpdate": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TeamMemberUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}